+++ chatroom/settingdialog.cc
@@ -1,8 +1,14 @@
 #include "settingdialog.hpp"
 
+#include <QJsonDocument>
+#include <QJsonObject>
+#include <QKeyEvent>
 #include <QShortcut>
 
+#include "global.hpp"
+#include "tcpmanager.hpp"
 #include "ui_settingdialog.h"
+#include "userinfo.hpp"
 
 SettingDialog::SettingDialog(QWidget* parent) : QDialog(parent), ui(new Ui::SettingDialog)
 {
@@ -16,4 +22,17 @@ SettingDialog::SettingDialog(QWidget* parent) : QDialog(parent), ui(new Ui::Sett
 SettingDialog::~SettingDialog()
 {
   delete ui;
+}
+
+// NOLINTNEXTLINE(readability-convert-member-functions-to-static)
+void SettingDialog::on_exit_button_clicked()
+{
+  // 发送退出登录请求
+  QJsonObject obj;
+  obj["uuid"] = UserInfo::GetInstance().GetUUID();
+
+  QJsonDocument doc{obj};
+  QString jsonStr = doc.toJson(QJsonDocument::Compact);
+
+  TcpManager::GetInstance().sig_send_data(ReqID::ID_EXIT_LOGIN, jsonStr);
 }