+++ services/ChatServer/src/utils/grpc/service/chat_service.cc
@@ -1,0 +1,6 @@
+#include "chat_service.hpp"
+
+namespace utils
+{
+
+}  // namespace utils